pyproject.toml
[project]
name = "cafedatacollection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-slugify>=8.0.4",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
